Use xdmf instead of xml for file I/O

+++ src/test.py
@@ -7,26 +7,24 @@ import matplotlib.pyplot as plt
 import glob
 import os
 import shutil
+import time
+from mpi4py import MPI
 from functools import partial
 import dolfin_adjoint as da
 from pyadjoint.overloaded_type import create_overloaded_object
-# import moola
-# import scipy.optimize as opt
 from . import arguments
 from .constitutive import *
 
-
-fe.parameters["form_compiler"]["quadrature_degree"] = 4
+rank = MPI.COMM_WORLD.Get_rank()
 
 
 class PDE(object):
-    def __init__(self, problem='inverse'):
+    def __init__(self, problem):
         self.case_name = "brittle"
-        self.displacements = 1e-1*np.linspace(0, 0.15, 100)
+        self.displacements = 1e-1*np.linspace(0, 0.1, 150)
         self.relaxation_parameters =  np.linspace(1, 1, len(self.displacements))
-        self.staggered_tol = 1e-5
-        self.staggered_maxiter = 1000 
-        self.display_intermediate_results = False
+        # self.staggered_tol = 1e-5
+        # self.staggered_maxiter = 1000
         self.delta_u_recorded = []
         self.sigma_recorded = []     
         self.problem = problem 
@@ -40,22 +38,26 @@ class PDE(object):
             self.adjoint_optimization()
         elif self.problem == 'forward':
             self.opt_step = opt_step
-            self.build_mesh(f'data/xml/{self.case_name}/inverse/mesh_{self.opt_step}.xml')
+            self.build_mesh(f'data/xdmf/{self.case_name}/inverse/mesh_{self.opt_step}.xdmf')
             self.staggered_solve()
         elif self.problem == 'post-processing':
             self.opt_step = opt_step
             self.plot_force_displacement()
             plt.ioff()
             plt.show()
+        elif self.problem == 'debug':
+            self.opt_step = opt_step
+            self.build_mesh(f'data/xdmf/{self.case_name}/inverse/mesh_{self.opt_step}.xdmf')
+            self.staggered_solve()
         else:
             raise ValueError('Unknown problem mode!')
 
 
     def preparation(self):
-        if self.problem == 'inverse':
+        if self.problem == 'inverse' and rank == 0:
             print(f"\nDelete inverse problem related data...")
             shutil.rmtree(f'data/pvd/{self.case_name}/{self.problem}', ignore_errors=True)
-            shutil.rmtree(f'data/xml/{self.case_name}/{self.problem}', ignore_errors=True)
+            shutil.rmtree(f'data/xdmf/{self.case_name}/{self.problem}', ignore_errors=True)
 
 
     def build_mesh(self, mesh_file=None): 
@@ -70,8 +72,11 @@ class PDE(object):
         material_domain = plate - circle
         self.mesh = mshr.generate_mesh(material_domain, 50)
 
-        if self.problem == 'forward':
-            self.mesh = fe.Mesh(mesh_file)
+        if self.problem == 'forward' or self.problem == 'debug':
+            self.mesh = fe.Mesh()
+            xdmf_file_mesh = fe.XDMFFile(MPI.COMM_WORLD, mesh_file)
+            xdmf_file_mesh.read(self.mesh)
+            # self.mesh = fe.Mesh(mesh_file)
 
         # Add dolfin-adjoint dependency
         self.mesh  = create_overloaded_object(self.mesh)
@@ -131,15 +136,12 @@ class PDE(object):
 
 
     def mesh_deformation(self, h):
-
         zero = da.Constant([0.] * self.mesh.geometric_dimension())
-
         S = fe.VectorFunctionSpace(self.mesh, "CG", 1)
         s = da.Function(S, name="Mesh perturbation field")
         h_V = da.transfer_from_boundary(h, self.mesh)
         h_V.rename("Volume extension of h", "")
 
-        # Compute variable :math:`\mu`
         V = fe.FunctionSpace(self.mesh, "CG", 1)
         u, v = fe.TrialFunction(V), fe.TestFunction(V)
 
@@ -156,7 +158,6 @@ class PDE(object):
         mu = da.Function(V, name="mesh deformation mu")
         da.solve(a == l, mu, bcs=bcs)
 
-        # Compute the mesh deformation
         S = fe.VectorFunctionSpace(self.mesh, "CG", 1)
         u, v = fe.TrialFunction(S), fe.TestFunction(S)
 
@@ -236,17 +237,18 @@ class PDE(object):
         self.solver_u = da.NonlinearVariationalSolver(p_u)
         self.solver_d  = da.NonlinearVariationalSolver(p_d)
 
-        if self.problem == 'forward':
-            vtkfile_u = fe.File(f'data/pvd/{self.case_name}/{self.problem}/step_{self.opt_step}/u.pvd')
-            vtkfile_d = fe.File(f'data/pvd/{self.case_name}/{self.problem}/step_{self.opt_step}/d.pvd')
+        if self.problem == 'forward' or self.problem == 'debug':
+            xdmf_file_sols = fe.XDMFFile(MPI.COMM_WORLD, f'data/xdmf/{self.case_name}/{self.problem}/step_{self.opt_step}/sols.xdmf')
+            xdmf_file_sols.parameters["functions_share_mesh"] = True
 
         self.J = 0
 
         for i, (disp, rp) in enumerate(zip(self.displacements, self.relaxation_parameters)):
-            print('\n')
-            print('=================================================================================')
-            print(f'>> Step {i}, disp boundary condition = {disp} [mm]')
-            print('=================================================================================')
+            if rank == 0:
+                print('\n')
+                print('=================================================================================')
+                print(f'>> Step {i}, disp boundary condition = {disp} [mm]')
+                print('=================================================================================')
 
             self.presLoad.t = disp
 
@@ -256,7 +258,9 @@ class PDE(object):
             newton_prm['relaxation_parameter'] = rp
  
             psi_plus = self.psi_plus(strain(fe.grad(self.x_new)))
+
             self.history.assign(da.project(fe.conditional(fe.gt(psi_plus, self.history), psi_plus, self.history), self.V))
+            # self.history.assign(da.project(psi_plus, self.V))
 
             self.solver_d.solve()
 
@@ -265,26 +269,25 @@ class PDE(object):
             # self.J += da.assemble(0.5 * self.d_new**2 * fe.dx)
             self.J -= da.assemble(self.sigma[1, 1]*self.ds(1))
 
-            if self.problem == 'forward':
-                vtkfile_u << self.x_new
-                vtkfile_d << self.d_new
+            if self.problem == 'forward' or self.problem == 'debug':
+                xdmf_file_sols.write(self.x_new, i)
+                xdmf_file_sols.write(self.d_new, i)
 
             force_upper = float(da.assemble(self.sigma[1, 1]*self.ds(1)))
-            print(f"Force upper {force_upper}")
+            if rank == 0:
+                print(f"Force upper {force_upper}")
             self.delta_u_recorded.append(disp)
             self.sigma_recorded.append(force_upper)
 
-            # if self.display_intermediate_results and i % 20 == 0:
-            #     self.plot_force_displacement()
 
-        alpha = 1e1
+        alpha = 1e2
         Vol = da.assemble(self.one * fe.dx(domain=self.mesh))
         self.J += alpha * ((self.length * self.height - Vol) - self.Vol0)**2
      
         (x, y) = fe.SpatialCoordinate(self.mesh)
         Bc1 = (self.length**2 * self.height / 2 - da.assemble(x * fe.dx(domain=self.mesh))) / (self.length * self.height - Vol)
         Bc2 = (self.length * self.height**2 / 2 - da.assemble(y * fe.dx(domain=self.mesh))) / (self.length * self.height - Vol)
-        beta = 1e1
+        beta = 1e2
         self.J += beta * ((Bc1 - self.xcenter)**2 + (Bc2 - self.ycenter)**2)
 
         if self.problem == 'forward':
@@ -299,27 +302,25 @@ class PDE(object):
 
         def eval_callback(j, m):
             '''
-            Hint: m has the same type as self.h
+            Hint: m has the same type as self.h 
+            Might be useful.
             '''
             # print(f"Parameter is {np.array(m.vector())}")
             print(f"Objective is {j}")
             self.object_values.append(j)
 
+
         control = da.Control(self.h)
         Jhat = reduced_functional = da.ReducedFunctional(self.J, control, eval_cb_post=eval_callback)
-
-        # vtkfile_u_opt = fe.File(f'data/pvd/{self.case_name}/u_opt.pvd')
-        # vtkfile_d_opt = fe.File(f'data/pvd/{self.case_name}/d_opt.pvd')
 
         def iteration_callback(x):
             ''' Callback function for optimization
 
             Parameter
             ---------
-            x is a 1D numpy array
+            x: 1D numpy array
             '''
             iteration_callback.count += 1
-            print("BFGS call back...")
             mesh_copy = fe.Mesh(self.mesh_initial)
             h_callback = fe.Function(self.S_b)
             h_callback.vector()[:] = x
@@ -327,37 +328,19 @@ class PDE(object):
             fe.ALE.move(mesh_copy, s)
             # self.plot_mesh(mesh_copy)
             vtkfile_mesh << mesh_copy
-            xml_mesh = fe.File(f'data/xml/{self.case_name}/{self.problem}/mesh_{iteration_callback.count}.xml')
-            xml_mesh << mesh_copy
-            
-            print(f"max h {np.max(np.absolute(x))}")
+            xdmf_file_mesh = fe.XDMFFile(MPI.COMM_WORLD, f'data/xdmf/{self.case_name}/{self.problem}/mesh_{iteration_callback.count}.xdmf')
+            xdmf_file_mesh.write(mesh_copy)
+            print(f"BFGS call back...:max h {np.max(np.absolute(x))}")
 
 
         iteration_callback.count = 0
+        xdmf_file_mesh = fe.XDMFFile(MPI.COMM_WORLD, f'data/xdmf/{self.case_name}/{self.problem}/mesh_{iteration_callback.count}.xdmf')
+        xdmf_file_mesh.write(self.mesh)
         vtkfile_mesh = fe.File(f'data/pvd/{self.case_name}/{self.problem}/mesh.pvd')
         vtkfile_mesh << self.mesh
-        xml_mesh = fe.File(f'data/xml/{self.case_name}/{self.problem}/mesh_{iteration_callback.count}.xml')
-        xml_mesh << self.mesh
 
         h_opt = da.minimize(reduced_functional, method="L-BFGS-B", tol=1e-20, bounds=(-2., 2.), callback=iteration_callback,
-            options={"disp": True, "maxiter": 10})
-
-        # np.save(f'data/numpy/{self.case_name}/object_values.npy', np.array(self, object_values))
-
-        # s = self.mesh_deformation(h_opt)
-        # fe.ALE.move(self.mesh, s)
-        # self.d_new.vector()[:] = 0
-        # self.x_new.vector()[:] = 0
-        # for i, (disp, rp) in enumerate(zip(self.displacements, self.relaxation_parameters)):
-        #     self.presLoad.t = disp
-        #     self.solver_d.solve()
-        #     self.solver_u.solve()
-        #     vtkfile_u_opt << self.x_new
-        #     vtkfile_d_opt << self.d_new
-
-        # print(self.object_values)
-
-        # self.show_optimization_progress()
+            options={"disp": True, "maxiter": 2})
 
 
     def plot_mesh(self, mesh_copy):
@@ -414,7 +397,15 @@ def plot_comparison(case_name):
     plt.show()
 
 
-def test(args):
+def main(args):
+    begin_time = time.time()
+    for i in range(3):
+        pde = PDE('debug')
+        pde.run(i)
+    end_time = time.time()
+    if rank == 0:
+        print(f"Wall time elapsed: {end_time-begin_time}")
+
     # pde = PDE('inverse')
     # pde.run()
 
@@ -422,13 +413,12 @@ def test(args):
     #     pde = PDE('forward')
     #     pde.run(i)
 
-    for i in range(3):
-        pde = PDE('post-processing')
-        pde.run(i)
-
+    # for i in range(3):
+    #     pde = PDE('post-processing')
+    #     pde.run(i)
 
 
 if __name__ == '__main__':
     args = arguments.args
-    # test(args)
-    plot_comparison('brittle')
+    main(args)
+    # plot_comparison('brittle')
